# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Masafah API Documentation",
        "description": "Api description...",
        "version": "1.0.0"
    },
    "paths": {
...
        "/~tvavisa/masafah/public/api/company/getProfile": {
            "get": {
                "tags": [
                    "Company Profile"
                ],
                "summary": "Get Company Profile",
                "operationId": "getProfile",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Shipment not found"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/company/updateProfile": {
            "put": {
                "tags": [
                    "Company Profile"
                ],
                "summary": "Update Company profile",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Update profile body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "description": "Company name - *(Required)",
                                    "type": "string",
                                    "example": "Aramex"
                                },
                                "email": {
                                    "description": "Company email - *(Required)",
                                    "type": "string",
                                    "example": "info@aramex.com"
                                },
                                "country_id": {
                                    "description": "Company country ID - *(Required)",
                                    "type": "integer",
                                    "example": 4
                                },
                                "image": {
                                    "description": "Profile image base64",
                                    "type": "string",
                                    "example": "9vjklhtyi9765/87997jhbsdfh/iutvs......"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/company/changeMobileNumber": {
            "patch": {
                "tags": [
                    "Company Profile"
                ],
                "summary": "Change Company's Mobile number",
                "operationId": "changeMobileNumber",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Change number body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "mobile": {
                                    "description": "Company Mobile number - *(Required)",
                                    "type": "string",
                                    "example": "66341897"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "409": {
                        "description": "Mobile number already registered"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/company/updateMobileNumber": {
            "patch": {
                "tags": [
                    "Company Profile"
                ],
                "summary": "Update Company's Mobile number",
                "operationId": "updateMobileNumber",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Change number body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "mobile": {
                                    "description": "Company Mobile number - *(Required)",
                                    "type": "string",
                                    "example": "66341897"
                                },
                                "otp": {
                                    "description": "Received OTP - *(Required)",
                                    "type": "string",
                                    "example": "46137"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "401": {
                        "description": "Invalid OTP. Unauthorized"
                    },
                    "409": {
                        "description": "Mobile number already registered"
                    }
                }
            }
        },
...
        "/~tvavisa/masafah/public/api/user/getProfile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get User Profile",
                "operationId": "getUserProfile",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/user/updateProfile": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update User profile",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Update profile body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "fullname": {
                                    "description": "Users full name",
                                    "type": "string",
                                    "example": "Fakhruddin Tahery"
                                },
                                "email": {
                                    "description": "User email",
                                    "type": "string",
                                    "example": "ftahery@vavisa-kw.com"
                                },
                                "image": {
                                    "description": "Profile image base64",
                                    "type": "string",
                                    "example": "9vjklhtyi9765/87997jhbsdfh/iutvs......"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/user/changeMobileNumber": {
            "patch": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Change User's Mobile number",
                "operationId": "changeMobileNumber",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Change number body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "mobile": {
                                    "description": "User Mobile number - *(Required)",
                                    "type": "string",
                                    "example": "66341897"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "409": {
                        "description": "Mobile number already registered"
                    }
                }
            }
        },
        "/~tvavisa/masafah/public/api/user/updateMobileNumber": {
            "patch": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update User's Mobile number",
                "operationId": "updateMobileNumber",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "user prefered language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "user access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Version",
                        "in": "header",
                        "description": "1.0.0",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Change number body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "mobile": {
                                    "description": "User's Mobile number - *(Required)",
                                    "type": "string",
                                    "example": "99653421"
                                },
                                "otp": {
                                    "description": "Received OTP - *(Required)",
                                    "type": "string",
                                    "example": "46137"
                                },
                                "type": {
                                    "description": "1-iOS, 2-Android - *(Required)",
                                    "type": "integer",
                                    "example": "46137"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "401": {
                        "description": "Invalid OTP. Unauthorized"
                    },
                    "409": {
                        "description": "Mobile number already registered"
                    }
                }
            }
        }
    },
...
}